+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 
 - Creating, editing or deleting lithological descriptions was not allowed for non-admin users.
+- Add missing translations when editing a locked borehole.
 
 ## v2.0.186 - 2023-02-13